ADDED USER VALIDATION IN ORDERS

+++ src/resolvers/OrderResolver.ts
@@ -11,8 +11,10 @@ import {
     Float,
 } from "type-graphql";
 import prisma from "../prisma/client";
-import { Order, OrderStatus } from "@generated/type-graphql";
+import { Order, OrderStatus, AccessRole } from "@generated/type-graphql";
 import { UserInputError } from "apollo-server-core";
+import { Context } from "../types";
+import { isUserAllowed } from "../utils/role";
 
 @InputType()
 export class ProductOrderCreateInput {
@@ -179,7 +181,15 @@ export class OrderResolver {
     async createOrder(
         @Arg("orderCreateInput", () => OrderCreateInput)
         orderCreateInput: OrderCreateInput,
+        @Ctx() ctx: Context,
     ) {
+        if (
+            !isUserAllowed(ctx.user!.role, [
+                AccessRole.orders,
+                AccessRole.admin,
+            ])
+        )
+            throw new UserInputError("Not Authorized");
         const orderCanBeCreated =
             await OrderResolver.checkRawMaterialAvailability(orderCreateInput);
         if (!orderCanBeCreated)
@@ -235,7 +245,15 @@ export class OrderResolver {
     async updateOrderDetails(
         @Arg("id", () => Int) id: number,
         @Arg("data", () => OrderUpdateInput) data: OrderUpdateInput,
+        @Ctx() ctx: Context,
     ) {
+        if (
+            !isUserAllowed(ctx.user!.role, [
+                AccessRole.orders,
+                AccessRole.admin,
+            ])
+        )
+            throw new UserInputError("Not Authorized");
         const order = await prisma.order.update({
             where: {
                 id,
@@ -264,7 +282,15 @@ export class OrderResolver {
         @Arg("id", () => Int) id: number,
         @Arg("productOrderCreateInput", () => [ProductOrderCreateInput])
         productOrderCreateInput: ProductOrderCreateInput[],
+        @Ctx() ctx: Context,
     ) {
+        if (
+            !isUserAllowed(ctx.user!.role, [
+                AccessRole.orders,
+                AccessRole.admin,
+            ])
+        )
+            throw new UserInputError("Not Authorized");
         const order = await prisma.order.update({
             where: {
                 id,
@@ -281,7 +307,17 @@ export class OrderResolver {
 
     @Mutation(() => Order)
     @Authorized()
-    async removeProductFromOrder(@Arg("id", () => Int) id: number) {
+    async removeProductFromOrder(
+        @Arg("id", () => Int) id: number,
+        @Ctx() ctx: Context,
+    ) {
+        if (
+            !isUserAllowed(ctx.user!.role, [
+                AccessRole.orders,
+                AccessRole.admin,
+            ])
+        )
+            throw new UserInputError("Not Authorized");
         return await prisma.productOrder.delete({
             where: {
                 id,
@@ -294,7 +330,15 @@ export class OrderResolver {
     async updateStatus(
         @Arg("id", () => Int) id: number,
         @Arg("status", () => Int) status: OrderStatus,
+        @Ctx() ctx: Context,
     ) {
+        if (
+            !isUserAllowed(ctx.user!.role, [
+                AccessRole.orders,
+                AccessRole.admin,
+            ])
+        )
+            throw new UserInputError("Not Authorized");
         try {
             const order = prisma.order.update({
                 where: {
@@ -312,7 +356,14 @@ export class OrderResolver {
 
     @Mutation(() => Order)
     @Authorized()
-    async deleteOrder(@Arg("id", () => Int) id: number) {
+    async deleteOrder(@Arg("id", () => Int) id: number, @Ctx() ctx: Context) {
+        if (
+            !isUserAllowed(ctx.user!.role, [
+                AccessRole.orders,
+                AccessRole.admin,
+            ])
+        )
+            throw new UserInputError("Not Authorized");
         try {
             const order = prisma.order.update({
                 where: {
